Remove / move utils
