+++ inlang/source-code/paraglide/paraglide-js/src/cli/commands/init/command.ts
@@ -7,7 +7,7 @@ import { telemetry } from "~/services/telemetry/implementation.js"
 import { Logger } from "~/services/logger/index.js"
 import { findRepoRoot, openRepository, type Repository } from "@lix-js/client"
 import { findPackageJson } from "~/services/environment/package.js"
-import { promptSelection } from "./utils.js"
+import { promptSelection } from "~/cli/utils.js"
 import { checkForUncommittedChanges } from "~/cli/steps/check-for-uncomitted-changes.js"
 import { initializeInlangProject } from "~/cli/steps/initialize-inlang-project.js"
 import { maybeAddSherlock } from "~/cli/steps/maybe-add-sherlock.js"
