+++ inlang/source-code/paraglide/paraglide-js/src/cli/steps/maybe-add-sherlock.ts
@@ -1,5 +1,5 @@
 import type { Logger } from "~/services/logger/index.js"
-import { prompt } from "~/cli/commands/init/utils.js"
+import { prompt } from "~/cli/utils.js"
 import type { Repository } from "@lix-js/client"
 import type { CliStep } from "../utils.js"
 import type { InlangProject } from "@inlang/sdk"
