+++ inlang/source-code/paraglide/paraglide-js/src/cli/steps/update-ts-config.ts
@@ -1,7 +1,7 @@
 import type { Repository } from "@lix-js/client"
 import type { CliStep } from "../utils.js"
 import type { Logger } from "~/services/logger/index.js"
-import { prompt } from "~/cli/commands/init/utils.js"
+import { prompt } from "~/cli/utils.js"
 import JSON5 from "json5"
 import { pathExists } from "~/services/file-handling/exists.js"
 import nodePath from "node:path"
