+++ inlang/source-code/paraglide/paraglide-js/src/cli/utils.ts
@@ -1,3 +1,5 @@
+import consola from "consola"
+
 /**
  * One step in a CLI chain.
  * Defines which types the context needs to extend and how it extends the context
@@ -22,3 +24,21 @@
  * ```
  */
 export type CliStep<In extends object, Out> = <Ctx extends In>(ctx: Ctx) => Promise<Ctx & Out>
+
+/**
+ * Wrapper to exit the process if the user presses CTRL+C.
+ */
+export const prompt: typeof consola.prompt = async (message, options) => {
+	const response = await consola.prompt(message, options)
+	if (response?.toString() === "Symbol(clack:cancel)") {
+		process.exit(0)
+	}
+	return response
+}
+
+export const promptSelection = async <T extends string>(
+	message: string,
+	options: { initial?: T; options: { label: string; value: T }[] } = { options: [] }
+): Promise<T> => {
+	return prompt(message, { type: "select", ...options }) as unknown as Promise<T>
+}
